pyproject.toml
[tool.poetry]
name = "slack-leaderboard"
version = "0.0.0"
description = ""
authors = ["Milo Weinberg <iapetus011@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
slack-bolt = "^1.15.5"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"
tortoise-orm = {extras = ["asyncpg"], version = "^0.19.2"}
asyncpg = "^0.27.0"
fastapi = "^0.88.0"
uvicorn = "^0.20.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
flake8 = "^6.0.0"
isort = "^5.10.1"
black = "^22.10.0"
autoflake = "^2.0.0"
aerich = "^0.7.1"

[tool.aerich]
tortoise_orm = "src.config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
plugins = ["pydantic.mypy"]
files = ["src/**/*.py"]
python_version = "3.10"
warn_unused_configs = true
namespace_packages = true
explicit_package_bases = true
